requirements.txt
python-telegram-bot==22.6
psutil==5.9.8
pyautogui==0.9.54
screen-brightness-control==0.23.0
requests==2.31.0
rapidfuzz
pywinauto
pygetwindow
